move config into separate package
