+++ cmd/root.go
@@ -5,6 +5,7 @@ import (
 	"sort"
 
 	"github.com/rebuy-de/aws-nuke/pkg/awsutil"
+	"github.com/rebuy-de/aws-nuke/pkg/config"
 	"github.com/rebuy-de/aws-nuke/resources"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -52,7 +53,7 @@ func NewRootCommand() *cobra.Command {
 
 		n := NewNuke(params, *account)
 
-		n.Config, err = LoadConfig(n.Parameters.ConfigPath)
+		n.Config, err = config.LoadConfig(n.Parameters.ConfigPath)
 		if err != nil {
 			return err
 		}
