+++ pkg/config/config.go
@@ -1,4 +1,4 @@
-package cmd
+package config
 
 import (
 	"fmt"
